# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Web API",
        "contact": {
            "name": "Enido Vasku",
            "email": "enidovasku@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
        "/api/v1/basketitems/": {
            "post": {
                "tags": [
                    "Add Product"
                ],
                "operationId": "App\\Http\\Controllers\\BasketItemsController::insert",
                "parameters": [
                    {
                        "name": "productId",
                        "in": "query",
                        "description": "Name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "quantity",
                        "in": "query",
                        "description": "Description",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "price",
                        "in": "query",
                        "description": "Price",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "userId",
                        "in": "query",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns some sample category things",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. When required parameters were not supplied."
                    }
                }
            }
        },
        "/api/v1/basketitems/{id}": {
            "get": {
                "tags": [
                    "Basket"
                ],
                "operationId": "App\\Http\\Controllers\\BasketItemsController::index",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "User Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns some sample category things",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. When required parameters were not supplied."
                    }
                }
            }
        },
        "/api/v1/product/": {
            "get": {
                "tags": [
                    "Products"
                ],
                "operationId": "App\\Http\\Controllers\\ProductController::index",
                "responses": {
                    "200": {
                        "description": "Returns some sample category things",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. When required parameters were not supplied."
                    }
                }
            },
            "post": {
                "tags": [
                    "Create Product"
                ],
                "operationId": "App\\Http\\Controllers\\ProductController::create",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Description",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "price",
                        "in": "query",
                        "description": "Price",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "quantity",
                        "in": "query",
                        "description": "Quantity",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns some sample category things",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. When required parameters were not supplied."
                    }
                }
            }
        }
    }
}